Fix image handling for channel messages and refactor tool enhancement
- Enable base64 image conversion for user messages so AI can see images
  from Telegram/Slack/WhatsApp instead of hallucinating URLs
- Extract tool enhancement functions to lib/messages/tool-enhancement.ts
  to comply with Next.js route export restrictions
- Fix type errors in reset-channel route and delivery module
